+++ app/src/main/java/com/sopt/now/compose/presentation/ui/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.presentation.main
+package com.sopt.now.compose.presentation.ui.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -9,9 +9,9 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
-import com.sopt.now.compose.presentation.auth.navigation.AuthNavigator
-import com.sopt.now.compose.presentation.home.navigation.HomeNavigator
-import com.sopt.now.compose.presentation.navigator.MainNavHost
+import com.sopt.now.compose.presentation.ui.auth.navigation.AuthNavigator
+import com.sopt.now.compose.presentation.ui.home.navigation.HomeNavigator
+import com.sopt.now.compose.presentation.ui.navigator.MainNavHost
 import com.sopt.now.compose.ui.theme.CUSTOMTheme
 
 class MainActivity : ComponentActivity() {
@@ -27,7 +27,7 @@ class MainActivity : ComponentActivity() {
                     containerColor = MaterialTheme.colorScheme.background,
                     content = { paddingValues ->
                         MainNavHost(
-                            navController = navController,
+                            navHostController = navController,
                             modifier = Modifier.padding(paddingValues),
                             authNavigator = authNavigator,
                             homeNavigator = homeNavigator
